pyproject.toml
[project]
name = "norreroute"
version = "0.3.0"
description = "A pragmatic, provider-agnostic Python library for calling LLMs"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Leo", email = "leo@talenddev.com"}
]
dependencies = [
    "httpx>=0.27",
    "anthropic>=0.28",
    "pydantic-settings>=2.3",
]
requires-python = ">=3.12"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/norreroute"]

[project.optional-dependencies]
otel = ["opentelemetry-api>=1.25"]

[dependency-groups]
dev = [
    "pytest>=8",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5",
    "pytest-mock>=3.15.1",
    "respx>=0.21",
    "ruff>=0.4",
    "mypy>=1.10",
    "anyio>=4",
    "pip-audit>=2.10.0",
    "opentelemetry-sdk>=1.25",
    "opentelemetry-api>=1.25",
]

[tool.ruff]
target-version = "py312"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["B017"]

[tool.ruff.lint.isort]
known-first-party = ["norreroute"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
packages = ["norreroute"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "--cov=src --cov-report=term-missing"

# Allow running with zero tests collected (scaffold phase)
empty_parameter_set_mark = "skip"
